Move signal handling into its own translation unit to limit overlap
